# api/wxapp/post_api.py
"""
微信小程序帖子API
提供帖子管理相关的API接口
"""
from fastapi import HTTPException, Path as PathParam, Depends, Query
from pydantic import BaseModel, Field, validator, root_validator
from typing import Dict, Any, Optional, List
import json
import traceback
from datetime import datetime

# 导入通用组件
from api.common import get_api_logger, handle_api_errors, create_standard_response
from api import wxapp_router as router
from api.wxapp.common_utils import format_datetime, prepare_db_data, process_json_fields

# 导入数据库操作函数
from etl.load.py_mysql import (
    insert_record, update_record, delete_record, 
    query_records, count_records, get_record_by_id,
    execute_raw_query
)

# 帖子模型
class PostBase(BaseModel):
    """帖子基础信息"""
    openid: str = Field(..., description="发布用户openid")
    title: str = Field(..., description="帖子标题", min_length=1, max_length=100)
    content: str = Field(..., description="帖子内容")
    images: Optional[List[str]] = Field(default=[], description="图片URL列表")
    tags: Optional[List[str]] = Field(default=[], description="标签列表")
    category_id: Optional[int] = Field(None, description="分类ID")
    location: Optional[str] = Field(None, description="位置信息")
    nick_name: Optional[str] = Field(None, description="用户名称")
    avatar: Optional[str] = Field(None, description="用户头像URL")
    
    @validator('openid')
    def validate_openid(cls, v):
        """确保openid不为空"""
        if not v:
            raise ValueError("用户openid不能为空")
        return v
    
    @validator('title')
    def validate_title(cls, v):
        if not v or not v.strip():
            raise ValueError("帖子标题不能为空")
        return v.strip()
    
    @validator('content')
    def validate_content(cls, v):
        if not v or not v.strip():
            raise ValueError("帖子内容不能为空")
        return v.strip()

class PostCreate(PostBase):
    """创建帖子请求"""
    pass

class PostUpdate(BaseModel):
    """更新帖子请求"""
    title: Optional[str] = None
    content: Optional[str] = None
    images: Optional[List[str]] = None
    tags: Optional[List[str]] = None
    category_id: Optional[int] = None
    location: Optional[str] = None
    status: Optional[int] = None
    extra: Optional[Dict[str, Any]] = None
    
    @validator('title')
    def validate_title(cls, v):
        if v is not None and not v.strip():
            raise ValueError("帖子标题不能为空")
        return v.strip() if v else None
    
    @validator('content')
    def validate_content(cls, v):
        if v is not None and not v.strip():
            raise ValueError("帖子内容不能为空")
        return v.strip() if v else None

class PostResponse(PostBase):
    """帖子响应"""
    id: int
    create_time: str
    update_time: str
    view_count: int = 0
    like_count: int = 0
    comment_count: int = 0
    favorite_count: int = 0
    liked_users: List[str] = []
    favorite_users: List[str] = []
    status: int = 1
    platform: str = "wxapp"
    is_deleted: int = 0
    extra: Optional[Dict[str, Any]] = None

# API端点
@router.post("/posts", response_model=Dict[str, Any], summary="创建帖子")
@handle_api_errors("创建帖子")
async def create_post(
    post: PostCreate,
    api_logger=Depends(get_api_logger)
):
    """创建新帖子"""
    # 准备数据
    post_data = post.dict()
    
    # 记录请求详情
    api_logger.debug(f"创建帖子请求数据: {post_data}")
    
    # 处理JSON字段
    if 'images' in post_data and post_data['images']:
        post_data['images'] = json.dumps(post_data['images'])
    else:
        post_data['images'] = json.dumps([])
        
    if 'tags' in post_data and post_data['tags']:
        post_data['tags'] = json.dumps(post_data['tags'])
    else:
        post_data['tags'] = json.dumps([])
    
    # 初始化点赞和收藏用户
    post_data['liked_users'] = json.dumps([])
    post_data['favorite_users'] = json.dumps([])
    
    # 添加其他默认值
    post_data['view_count'] = 0
    post_data['like_count'] = 0
    post_data['comment_count'] = 0
    post_data['favorite_count'] = 0  # 添加收藏数量初始化
    post_data['status'] = 1
    post_data['is_deleted'] = 0
    post_data['platform'] = 'wxapp'
    post_data['extra'] = json.dumps({})
    
    # 准备数据库数据
    post_data = prepare_db_data(post_data, is_create=True)
    
    # 将字段名映射到新的表结构
    if 'user_id' in post_data:
        post_data['openid'] = post_data.pop('user_id')
    if 'author_name' in post_data:
        post_data['nick_name'] = post_data.pop('author_name')
    if 'author_avatar' in post_data:
        post_data['avatar'] = post_data.pop('author_avatar')
    if 'user_name' in post_data:
        post_data['nick_name'] = post_data.pop('user_name')
    if 'user_avatar' in post_data:
        post_data['avatar'] = post_data.pop('user_avatar')
    
    # 移除数据库中不存在的字段
    if 'wxapp_id' in post_data:
        post_data.pop('wxapp_id', None)
    
    # 创建帖子
    try:
        post_id = insert_record('wxapp_posts', post_data)
        if not post_id:
            api_logger.error("插入帖子失败")
            raise HTTPException(status_code=500, detail="创建帖子失败")
        
        # 获取新创建的帖子
        new_post = get_record_by_id('wxapp_posts', post_id)
        if not new_post:
            api_logger.error(f"无法获取新创建的帖子: ID={post_id}")
            raise HTTPException(status_code=500, detail="创建帖子后获取数据失败")
        
        # 处理JSON字段
        new_post = process_json_fields(new_post)
        
        api_logger.info(f"帖子创建成功: ID={post_id}")
        return create_standard_response(new_post)
    except Exception as e:
        if isinstance(e, HTTPException):
            raise e
        error_details = traceback.format_exc()
        api_logger.error(f"创建帖子异常: {str(e)}\n{error_details}")
        raise HTTPException(status_code=500, detail=f"创建帖子失败: {str(e)}")

@router.get("/posts/{post_id}", response_model=Dict[str, Any], summary="获取帖子详情")
@handle_api_errors("获取帖子")
async def get_post(
    post_id: int = PathParam(..., description="帖子ID"),
    update_view: bool = Query(True, description="是否更新浏览量"),
    api_logger=Depends(get_api_logger)
):
    """获取指定帖子详情"""
    post = get_record_by_id('wxapp_posts', post_id)
    if not post:
        raise HTTPException(status_code=404, detail="帖子不存在")
    
    # 处理JSON字段
    post = process_json_fields(post)
    
    # 更新浏览量
    if update_view:
        update_record('wxapp_posts', post_id, {
            'view_count': post['view_count'] + 1,
            'update_time': format_datetime(None)  # 使用当前时间
        })
        post['view_count'] += 1
    
    return create_standard_response(post)

@router.get("/posts", response_model=Dict[str, Any], summary="查询帖子列表")
@handle_api_errors("查询帖子列表")
async def list_posts(
    limit: int = Query(20, description="返回记录数量限制", ge=1, le=100),
    offset: int = Query(0, description="分页偏移量", ge=0),
    openid: Optional[str] = Query(None, description="按用户openid筛选"),
    category_id: Optional[int] = Query(None, description="按分类ID筛选"),
    tag: Optional[str] = Query(None, description="按标签筛选"),
    status: Optional[int] = Query(1, description="帖子状态: 1-正常, 0-禁用"),
    order_by: str = Query("update_time DESC", description="排序方式"),
    api_logger=Depends(get_api_logger)
):
    """获取帖子列表"""
    # 详细记录请求参数
    api_logger.info(f"Request: GET /posts with params: limit={limit}, offset={offset}, openid={openid}, category_id={category_id}, tag={tag}, status={status}, order_by={order_by}")
    
    try:
        # 构建查询条件
        conditions = {}
        if openid is not None:
            conditions['openid'] = openid
                
        if category_id is not None:
            conditions['category_id'] = category_id
            
        if status is not None:
            conditions['status'] = status
            
        # 添加未删除条件
        conditions['is_deleted'] = 0
        
        # 记录查询条件
        api_logger.debug(f"查询条件: {conditions}")
        
        # 查询帖子列表
        api_logger.debug(f"执行查询: 表=wxapp_posts, 条件={conditions}, 排序={order_by}, 限制={limit}, 偏移={offset}")
        posts = query_records(
            'wxapp_posts',
            conditions=conditions,
            order_by=order_by,
            limit=limit,
            offset=offset
        )
        api_logger.debug(f"查询结果数量: {len(posts)}")
        
        # 处理标签筛选（数据库级别无法直接处理JSON字段）
        if tag and posts:
            api_logger.debug(f"按标签筛选: {tag}")
            filtered_posts = []
            for post in posts:
                post = process_json_fields(post)
                if 'tags' in post and tag in post['tags']:
                    filtered_posts.append(post)
            posts = filtered_posts
            api_logger.debug(f"标签筛选后数量: {len(posts)}")
        else:
            # 处理所有帖子的JSON字段
            posts = [process_json_fields(post) for post in posts]
            
        # 获取总数
        total_count = count_records('wxapp_posts', conditions)
        api_logger.debug(f"总记录数: {total_count}")
        
        # 将结果转换为前端期望的格式
        api_logger.info(f"Response: 帖子列表查询成功，返回{len(posts)}条记录")
        
        # 返回标准响应，确保结果在data字段中
        return create_standard_response(posts)
    except Exception as e:
        error_details = traceback.format_exc()
        api_logger.error(f"Error: 查询帖子列表失败: {str(e)}\n{error_details}")
        # 重新抛出异常让handle_api_errors装饰器处理
        raise HTTPException(status_code=500, detail=f"查询帖子失败: {str(e)}")

@router.put("/posts/{post_id}", response_model=Dict[str, Any], summary="更新帖子")
@handle_api_errors("更新帖子")
async def update_post(
    post_update: PostUpdate,
    post_id: int = PathParam(..., description="帖子ID"),
    api_logger=Depends(get_api_logger)
):
    """更新帖子信息"""
    # 检查帖子是否存在
    post = get_record_by_id('wxapp_posts', post_id)
    if not post:
        raise HTTPException(status_code=404, detail="帖子不存在")
    
    # 过滤掉None值
    update_data = {k: v for k, v in post_update.dict().items() if v is not None}
    if not update_data:
        # 没有需要更新的字段，返回原帖子
        return create_standard_response(process_json_fields(post))
    
    # 处理JSON字段
    if 'images' in update_data:
        update_data['images'] = json.dumps(update_data['images'])
        
    if 'tags' in update_data:
        update_data['tags'] = json.dumps(update_data['tags'])
    
    # 添加更新时间
    update_data = prepare_db_data(update_data, is_create=False)
    
    # 更新记录
    success = update_record('wxapp_posts', post_id, update_data)
    if not success:
        raise HTTPException(status_code=500, detail="更新帖子失败")
    
    # 获取更新后的帖子
    updated_post = get_record_by_id('wxapp_posts', post_id)
    
    # 处理JSON字段
    updated_post = process_json_fields(updated_post)
    
    return create_standard_response(updated_post)

@router.delete("/posts/{post_id}", response_model=Dict[str, Any], summary="删除帖子")
@handle_api_errors("删除帖子")
async def delete_post(
    post_id: int = PathParam(..., description="帖子ID"),
    api_logger=Depends(get_api_logger)
):
    """删除帖子（标记删除）"""
    # 检查帖子是否存在
    post = get_record_by_id('wxapp_posts', post_id)
    if not post:
        raise HTTPException(status_code=404, detail="帖子不存在")
    
    # 标记删除
    success = update_record('wxapp_posts', post_id, {
        'is_deleted': 1,
        'status': 0,
        'update_time': format_datetime(None)  # 使用当前时间
    })
    
    if not success:
        raise HTTPException(status_code=500, detail="删除帖子失败")
    
    return create_standard_response({
        "success": True,
        "message": "帖子已删除"
    })

@router.post("/posts/{post_id}/like", response_model=Dict[str, Any], summary="点赞/取消点赞帖子")
@handle_api_errors("点赞帖子")
async def like_post(
    post_id: int = PathParam(..., description="帖子ID"),
    openid: str = Query(..., description="用户openid"),
    api_logger=Depends(get_api_logger)
):
    """点赞或取消点赞帖子"""
    # 检查帖子是否存在
    post = get_record_by_id('wxapp_posts', post_id)
    if not post:
        raise HTTPException(status_code=404, detail="帖子不存在")
    
    # 处理JSON字段
    post = process_json_fields(post)
    
    # 获取当前点赞用户列表
    liked_users = post.get('liked_users', [])
    
    # 判断是点赞还是取消点赞
    if openid in liked_users:
        # 取消点赞
        liked_users.remove(openid)
        like_count = max(0, post.get('like_count', 0) - 1)
        action = "取消点赞"
    else:
        # 点赞
        liked_users.append(openid)
        like_count = post.get('like_count', 0) + 1
        action = "点赞"
        
        # 如果是点赞操作，给帖子作者发送通知
        try:
            # 获取帖子作者openid
            author_openid = post.get('openid')
            # 如果点赞用户不是作者本人，才发送通知
            if author_openid and author_openid != openid:
                # 获取点赞用户信息
                openid_users = query_records(
                    'wxapp_users',
                    conditions={'openid': openid, 'is_deleted': 0}
                )
                if openid_users and len(openid_users) > 0:
                    liker_info = openid_users[0]
                    liker_name = liker_info.get('nick_name', '用户')
                    # 创建通知数据
                    notification_data = {
                        'user_id': author_openid,  # 通知接收者是帖子作者
                        'title': '收到新点赞',
                        'content': f"{liker_name} 点赞了你的帖子「{post.get('title', '无标题')}」",
                        'type': 'like',
                        'is_read': 0,
                        'sender_id': openid,  # 发送者是点赞用户
                        'related_id': str(post_id)  # 关联ID是帖子ID
                    }
                    # 准备数据
                    notification_data = prepare_db_data(notification_data, is_create=True)
                    # 插入通知记录
                    insert_record('wxapp_notifications', notification_data)
                    api_logger.debug(f"已为用户{author_openid}创建点赞通知")
        except Exception as e:
            # 通知创建失败不影响点赞功能
            api_logger.error(f"创建点赞通知失败: {str(e)}")
    
    # 获取当前时间
    current_time = datetime.now().strftime("%Y-%m-%d %H:%M:%S")
    
    # 更新帖子
    success = update_record('wxapp_posts', post_id, {
        'liked_users': json.dumps(liked_users),
        'like_count': like_count,
        'update_time': current_time  # 直接使用当前时间字符串
    })
    
    if not success:
        raise HTTPException(status_code=500, detail=f"{action}失败")
    
    return create_standard_response({
        "success": True,
        "message": f"{action}成功",
        "liked": openid in liked_users,
        "like_count": like_count
    })

@router.post("/posts/{post_id}/favorite", response_model=Dict[str, Any], summary="收藏/取消收藏帖子")
@handle_api_errors("收藏帖子")
async def favorite_post(
    post_id: int = PathParam(..., description="帖子ID"),
    openid: str = Query(..., description="用户openid"),
    is_favorite: bool = Query(True, description="是否收藏，默认为收藏操作"),
    api_logger=Depends(get_api_logger)
):
    """收藏或取消收藏帖子"""
    # 检查帖子是否存在
    post = get_record_by_id('wxapp_posts', post_id)
    if not post:
        raise HTTPException(status_code=404, detail="帖子不存在")
    
    # 处理JSON字段
    post = process_json_fields(post)
    
    # 获取当前收藏用户列表
    favorite_users = post.get('favorite_users', [])
    
    # 根据is_favorite参数决定是收藏还是取消收藏
    is_already_favorited = openid in favorite_users
    
    if is_already_favorited and not is_favorite:
        # 取消收藏
        favorite_users.remove(openid)
        action = "取消收藏"
        favorite_count = max(0, post.get('favorite_count', 0) - 1)  # 计算新的收藏数
    elif not is_already_favorited and is_favorite:
        # 收藏
        favorite_users.append(openid)
        action = "收藏"
        favorite_count = post.get('favorite_count', 0) + 1  # 计算新的收藏数
        
        # 如果是收藏操作，给帖子作者发送通知
        try:
            # 获取帖子作者openid
            author_openid = post.get('openid')
            # 如果收藏用户不是作者本人，才发送通知
            if author_openid and author_openid != openid:
                # 获取收藏用户信息
                openid_users = query_records(
                    'wxapp_users',
                    conditions={'openid': openid, 'is_deleted': 0}
                )
                if openid_users and len(openid_users) > 0:
                    favoriter_info = openid_users[0]
                    favoriter_name = favoriter_info.get('nick_name', '用户')
                    # 创建通知数据
                    notification_data = {
                        'user_id': author_openid,  # 通知接收者是帖子作者
                        'title': '收到新收藏',
                        'content': f"{favoriter_name} 收藏了你的帖子「{post.get('title', '无标题')}」",
                        'type': 'like',  # 使用like类型，因为UI上可能没有专门的收藏类型
                        'is_read': 0,
                        'sender_id': openid,  # 发送者是收藏用户
                        'related_id': str(post_id)  # 关联ID是帖子ID
                    }
                    # 准备数据
                    notification_data = prepare_db_data(notification_data, is_create=True)
                    # 插入通知记录
                    insert_record('wxapp_notifications', notification_data)
                    api_logger.debug(f"已为用户{author_openid}创建收藏通知")
        except Exception as e:
            # 通知创建失败不影响收藏功能
            api_logger.error(f"创建收藏通知失败: {str(e)}")
    else:
        # 状态已匹配，无需更改
        return create_standard_response({
            "success": True,
            "message": f"已经{'收藏' if is_favorite else '取消收藏'}",
            "favorited": is_favorite
        })
    
    # 获取当前时间
    current_time = datetime.now().strftime("%Y-%m-%d %H:%M:%S")
    
    # 更新帖子
    success = update_record('wxapp_posts', post_id, {
        'favorite_users': json.dumps(favorite_users),
        'favorite_count': favorite_count,  # 更新收藏数
        'update_time': current_time  # 直接使用当前时间字符串
    })
    
    if not success:
        raise HTTPException(status_code=500, detail=f"{action}失败")
    
    # 创建响应数据
    response_data = {
        "is_favorited": is_favorite,
        "favorite_count": favorite_count  # 返回最新收藏数量
    }
    
    # 返回标准响应
    return create_standard_response(
        data=response_data,
        message=f"{action}成功"
    ) 