# openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "nkuwiki API",
        "description": "\u5357\u5f00\u767e\u79d1\u77e5\u8bc6\u5e73\u53f0API\u670d\u52a1",
        "version": "1.0.0"
    },
    "paths": {
...
        "/wxapp/posts": {
            "post": {
                "tags": [
                    "wxapp"
                ],
                "summary": "\u521b\u5efa\u5e16\u5b50",
                "description": "\u521b\u5efa\u65b0\u5e16\u5b50",
                "operationId": "create_post_wxapp_posts_post",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PostCreate"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "title": "Response Create Post Wxapp Posts Post"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "get": {
                "tags": [
                    "wxapp"
                ],
                "summary": "\u67e5\u8be2\u5e16\u5b50\u5217\u8868",
                "description": "\u83b7\u53d6\u5e16\u5b50\u5217\u8868",
                "operationId": "list_posts_wxapp_posts_get",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "maximum": 100,
                            "minimum": 1,
                            "description": "\u8fd4\u56de\u8bb0\u5f55\u6570\u91cf\u9650\u5236",
                            "default": 20,
                            "title": "Limit"
                        },
                        "description": "\u8fd4\u56de\u8bb0\u5f55\u6570\u91cf\u9650\u5236"
                    },
                    {
                        "name": "offset",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "minimum": 0,
                            "description": "\u5206\u9875\u504f\u79fb\u91cf",
                            "default": 0,
                            "title": "Offset"
                        },
                        "description": "\u5206\u9875\u504f\u79fb\u91cf"
                    },
                    {
                        "name": "openid",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "description": "\u6309\u7528\u6237openid\u7b5b\u9009",
                            "title": "Openid"
                        },
                        "description": "\u6309\u7528\u6237openid\u7b5b\u9009"
                    },
                    {
                        "name": "category_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "integer"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "description": "\u6309\u5206\u7c7bID\u7b5b\u9009",
                            "title": "Category Id"
                        },
                        "description": "\u6309\u5206\u7c7bID\u7b5b\u9009"
                    },
                    {
                        "name": "tag",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "description": "\u6309\u6807\u7b7e\u7b5b\u9009",
                            "title": "Tag"
                        },
                        "description": "\u6309\u6807\u7b7e\u7b5b\u9009"
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "integer"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "description": "\u5e16\u5b50\u72b6\u6001: 1-\u6b63\u5e38, 0-\u7981\u7528",
                            "default": 1,
                            "title": "Status"
                        },
                        "description": "\u5e16\u5b50\u72b6\u6001: 1-\u6b63\u5e38, 0-\u7981\u7528"
                    },
                    {
                        "name": "order_by",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "description": "\u6392\u5e8f\u65b9\u5f0f",
                            "default": "update_time DESC",
                            "title": "Order By"
                        },
                        "description": "\u6392\u5e8f\u65b9\u5f0f"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "title": "Response List Posts Wxapp Posts Get"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/wxapp/posts/{post_id}": {
            "get": {
                "tags": [
                    "wxapp"
                ],
                "summary": "\u83b7\u53d6\u5e16\u5b50\u8be6\u60c5",
                "description": "\u83b7\u53d6\u6307\u5b9a\u5e16\u5b50\u8be6\u60c5",
                "operationId": "get_post_wxapp_posts__post_id__get",
                "parameters": [
                    {
                        "name": "post_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "description": "\u5e16\u5b50ID",
                            "title": "Post Id"
                        },
                        "description": "\u5e16\u5b50ID"
                    },
                    {
                        "name": "update_view",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "boolean",
                            "description": "\u662f\u5426\u66f4\u65b0\u6d4f\u89c8\u91cf",
                            "default": true,
                            "title": "Update View"
                        },
                        "description": "\u662f\u5426\u66f4\u65b0\u6d4f\u89c8\u91cf"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "title": "Response Get Post Wxapp Posts  Post Id  Get"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "wxapp"
                ],
                "summary": "\u66f4\u65b0\u5e16\u5b50",
                "description": "\u66f4\u65b0\u5e16\u5b50\u4fe1\u606f",
                "operationId": "update_post_wxapp_posts__post_id__put",
                "parameters": [
                    {
                        "name": "post_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "description": "\u5e16\u5b50ID",
                            "title": "Post Id"
                        },
                        "description": "\u5e16\u5b50ID"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PostUpdate"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "title": "Response Update Post Wxapp Posts  Post Id  Put"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "wxapp"
                ],
                "summary": "\u5220\u9664\u5e16\u5b50",
                "description": "\u5220\u9664\u5e16\u5b50\uff08\u6807\u8bb0\u5220\u9664\uff09",
                "operationId": "delete_post_wxapp_posts__post_id__delete",
                "parameters": [
                    {
                        "name": "post_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "description": "\u5e16\u5b50ID",
                            "title": "Post Id"
                        },
                        "description": "\u5e16\u5b50ID"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "title": "Response Delete Post Wxapp Posts  Post Id  Delete"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            }
        },
        "securitySchemes": {
            "HTTPBearer": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}